chore(tsc): typesafe 'sidenav' feature
