+++ src/components/sideNav/contextMenu.tsx
@@ -2,14 +2,27 @@ import React from 'react';
 import { connect } from 'react-redux';
 
 import { deleteNotes } from 'actions/noteActions';
+// FIXME
+interface SideNavContextMenuActions {
+	deleteNotes: any
+}
+
+interface SideNavContextMenuOwnProps {
+	selectedItems: Set<number>,
+	onDeleted(): void,
+	onSelectAll: React.ReactEventHandler<HTMLButtonElement>,
+	onDeselectAll: React.ReactEventHandler<HTMLButtonElement>
+};
+
+type SideNavContextMenuProps = SideNavContextMenuOwnProps & SideNavContextMenuActions;
 
 export function SideNavContextMenu({
-	selectedItems = null,
+	selectedItems,
 	deleteNotes,
 	onDeleted,
 	onSelectAll,
 	onDeselectAll
-}) {
+}: SideNavContextMenuProps) {
 	const show = selectedItems.size > 0;
 	const cxCssState = show ? ' active' : '';
 	const deleteCount = show ? ` (${selectedItems.size})` : '';
