+++ src/components/sideNav/floatButton.tsx
@@ -1,7 +1,11 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-export default function FloatButton({ label }) {
+interface FloatButtonProps {
+	label: string
+};
+
+export default function FloatButton({ label }: FloatButtonProps) {
 	return (
 		<Link to="/notes/new" className="btn-float">
 			{label}
