+++ src/components/sideNav/package.json
@@ -1,3 +1,3 @@
 {
-	"main": "SideNav.jsx"
+	"main": "SideNav.tsx"
 }
